frontend: forgot password validations

+++ backend/forgotPassword.py
@@ -15,12 +15,17 @@ class ForgotPassword:
 
         code  = self.__createCode__()
         
-        self.__saveCodeInDb__(code)
+        saved = self.__saveCodeInDb__(code)
 
-        email_manager.sendEmailToCreateNewPassword(email, code)
+        if saved:
+            email_manager.sendEmailToCreateNewPassword(email, code)
 
-        self.request.send_response(200, "OK")
-        self.request.end_headers()
+            self.request.send_response(200, "OK")
+            self.request.end_headers()
+        else:
+            self.request.send_error(500, "USER NOT FOUND")
+            self.request.end_headers()
+
 
     def __createCode__(self):
         code =  hashlib.sha256(str(random.randrange(9999999999)).encode()).hexdigest()
@@ -34,7 +39,12 @@ class ForgotPassword:
 
         user_id = db.getUserId(email)
 
+        if user_id == None:
+            return False
+
         db.deletePasswordRequest(user_id)
 
         db.addNewPasswordRequest(code, user_id)
+
+        return True
